allow value bin as oxford plot category

+++ src/genomehubs-api/src/api/v2/functions/attrTypes.js
@@ -13,7 +13,7 @@ const setProcessedType = (meta) => {
     return "float";
   }
   if (["long", "integer", "short", "byte"].includes(meta.type)) {
-    return "float";
+    return "integer";
   }
   if (meta.type == "keyword") {
     if (meta.summary) {
